physics and collision wip

+++ Arkanoid/include/collision.h
@@ -11,8 +11,7 @@ void draw_circle(const Circle& circle);
 struct AABB
 {
     int x_min, y_min, x_max, y_max;
-    static AABB make_from_position_size(float x, float y, float w, float h);
-    
+    static AABB make_from_position_size(int x, int y, int w, int h);
 };
 void draw_box(const AABB& box);
 bool aabb_intersect(const AABB& a, const AABB& b);
